ec2-script/requirements.txt
boto3==1.37.33
python-dotenv==1.1.0